fixed App Layout, routes and links in sidebar
