+++ src/App.jsx
@@ -2,20 +2,23 @@ import { Routes, Route } from "react-router-dom";
 import Home from "./pages/Home/Home.jsx";
 import About from "./pages/About.jsx";
 import NotFound from "./pages/NotFound.jsx";
-import Recipes from "./pages/Recipes.jsx";
+import RecipesPage from "./pages/Recipes/RecipesPage.jsx";
 import Recipe from "./pages/Recipe.jsx";
 import GenerateRecipe from "./pages/GenerateRecipe.jsx";
+import AppLayout from "./components/AppLayout.jsx";
 
 function App() {
   return (
     <>
       <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/generate-recipe" element={<GenerateRecipe />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/recipes" element={<Recipes />} />
-        <Route path="/recipes/:receipeId" element={<Recipe />} />
-        <Route path="*" element={<NotFound />} />
+        <Route element={<AppLayout />}>
+          <Route path="/" element={<Home />} />
+          <Route path="/generate-recipe" element={<GenerateRecipe />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/recipes" element={<RecipesPage />} />
+          <Route path="/recipes/:receipeId" element={<Recipe />} />
+          <Route path="*" element={<NotFound />} />
+        </Route>
       </Routes>
     </>
   );
