+++ src/pages/Recipes/RecipesPage.jsx
@@ -1,5 +1,7 @@
-export default function Recipes () {
+function RecipesPage () {
     return (
         <div>I am the Recipes Page with a list of all saved recipes.</div>
     );
 }
+
+export default RecipesPage;
